LayoutTests/svg/css/parse-length-expected.txt

CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="  100"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="100   "
CONSOLE MESSAGE: line 17: Error: Invalid negative value for <rect> attribute rx="-200px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="  100"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="100   "
CONSOLE MESSAGE: line 17: Error: Invalid negative value for <rect> attribute ry="-200px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="auto"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="  100"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="100   "
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="  100"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="100   "
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="  100"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="100   "
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="auto"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="100   px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="100px;"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="100px !important"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute rx="{ 100px }"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="auto"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="100   px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="100px;"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="100px !important"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute ry="{ 100px }"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="100   px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="100px;"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="100px !important"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute width="{ 100px }"
CONSOLE MESSAGE: line 17: Error: Invalid negative value for <rect> attribute width="-100px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="auto"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="100   px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="100px;"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="100px !important"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute x="{ 100px }"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="auto"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="100   px"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="100px;"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="100px !important"
CONSOLE MESSAGE: line 17: Error: Invalid value for <rect> attribute y="{ 100px }"
Test that 'width' presentation attribute is parsed with CSS presentation rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle("cx", "  100") is "100px"
PASS computedStyle("cx", "100   ") is "100px"
PASS computedStyle("cx", "100px") is "100px"
PASS computedStyle("cx", "1em") is "16px"
PASS computedStyle("cx", "1ex") is "12.800000190734863px"
PASS computedStyle("cx", "20%") is "20%"
PASS computedStyle("cx", "-200px") is "-200px"
PASS computedStyle("cy", "  100") is "100px"
PASS computedStyle("cy", "100   ") is "100px"
PASS computedStyle("cy", "100px") is "100px"
PASS computedStyle("cy", "1em") is "16px"
PASS computedStyle("cy", "1ex") is "12.800000190734863px"
PASS computedStyle("cy", "20%") is "20%"
PASS computedStyle("cy", "-200px") is "-200px"
PASS computedStyle("r", "  100") is "100px"
PASS computedStyle("r", "100   ") is "100px"
PASS computedStyle("r", "100px") is "100px"
PASS computedStyle("r", "1em") is "16px"
PASS computedStyle("r", "1ex") is "12.800000190734863px"
PASS computedStyle("r", "20%") is "20%"
PASS computedStyle("r", "-200px") is "-200px"
PASS computedStyle("rx", "  100") is "100px"
PASS computedStyle("rx", "100   ") is "100px"
PASS computedStyle("rx", "100px") is "100px"
PASS computedStyle("rx", "1em") is "16px"
PASS computedStyle("rx", "1ex") is "12.800000190734863px"
PASS computedStyle("rx", "20%") is "20%"
PASS computedStyle("rx", "-200px") is "-200px"
PASS computedStyle("ry", "  100") is "100px"
PASS computedStyle("ry", "100   ") is "100px"
PASS computedStyle("ry", "100px") is "100px"
PASS computedStyle("ry", "1em") is "16px"
PASS computedStyle("ry", "1ex") is "12.800000190734863px"
PASS computedStyle("ry", "20%") is "20%"
PASS computedStyle("ry", "-200px") is "-200px"
PASS computedStyle("width", "auto") is "auto"
PASS computedStyle("width", "  100") is "100px"
PASS computedStyle("width", "100   ") is "100px"
PASS computedStyle("width", "100px") is "100px"
PASS computedStyle("width", "1em") is "16px"
PASS computedStyle("width", "1ex") is "12.800000190734863px"
PASS computedStyle("width", "20%") is "20%"
PASS computedStyle("x", "  100") is "100px"
PASS computedStyle("x", "100   ") is "100px"
PASS computedStyle("x", "100px") is "100px"
PASS computedStyle("x", "1em") is "16px"
PASS computedStyle("x", "1ex") is "12.800000190734863px"
PASS computedStyle("x", "20%") is "20%"
PASS computedStyle("x", "-200px") is "-200px"
PASS computedStyle("y", "  100") is "100px"
PASS computedStyle("y", "100   ") is "100px"
PASS computedStyle("y", "100px") is "100px"
PASS computedStyle("y", "1em") is "16px"
PASS computedStyle("y", "1ex") is "12.800000190734863px"
PASS computedStyle("y", "20%") is "20%"
PASS computedStyle("y", "-200px") is "-200px"
PASS computedStyle("cx", "auto") is "0px"
PASS computedStyle("cx", "100   px") is "0px"
PASS computedStyle("cx", "100px;") is "0px"
PASS computedStyle("cx", "100px !important") is "0px"
PASS computedStyle("cx", "{ 100px }") is "0px"
PASS computedStyle("cy", "auto") is "0px"
PASS computedStyle("cy", "100   px") is "0px"
PASS computedStyle("cy", "100px;") is "0px"
PASS computedStyle("cy", "100px !important") is "0px"
PASS computedStyle("cy", "{ 100px }") is "0px"
PASS computedStyle("r", "auto") is "0px"
PASS computedStyle("r", "100   px") is "0px"
PASS computedStyle("r", "100px;") is "0px"
PASS computedStyle("r", "100px !important") is "0px"
PASS computedStyle("r", "{ 100px }") is "0px"
PASS computedStyle("rx", "auto") is "0px"
PASS computedStyle("rx", "100   px") is "0px"
PASS computedStyle("rx", "100px;") is "0px"
PASS computedStyle("rx", "100px !important") is "0px"
PASS computedStyle("rx", "{ 100px }") is "0px"
PASS computedStyle("ry", "auto") is "0px"
PASS computedStyle("ry", "100   px") is "0px"
PASS computedStyle("ry", "100px;") is "0px"
PASS computedStyle("ry", "100px !important") is "0px"
PASS computedStyle("ry", "{ 100px }") is "0px"
PASS computedStyle("width", "100   px") is "auto"
PASS computedStyle("width", "100px;") is "auto"
PASS computedStyle("width", "100px !important") is "auto"
PASS computedStyle("width", "{ 100px }") is "auto"
PASS computedStyle("width", "-100px") is "auto"
PASS computedStyle("x", "auto") is "0px"
PASS computedStyle("x", "100   px") is "0px"
PASS computedStyle("x", "100px;") is "0px"
PASS computedStyle("x", "100px !important") is "0px"
PASS computedStyle("x", "{ 100px }") is "0px"
PASS computedStyle("y", "auto") is "0px"
PASS computedStyle("y", "100   px") is "0px"
PASS computedStyle("y", "100px;") is "0px"
PASS computedStyle("y", "100px !important") is "0px"
PASS computedStyle("y", "{ 100px }") is "0px"
PASS successfullyParsed is true

TEST COMPLETE
